chore: move resolvers out of typeDefs dir
